+++ src/main/java/pl/edu/agh/productivitypal/repository/CategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pl.edu.agh.productivitypal.model.Category;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 }
